move traverse_for_value to EngineResult
